+++ osf/models/mixins.py
@@ -36,10 +36,24 @@ from osf.utils import sanitize
 from osf.models.validators import validate_subject_hierarchy, validate_email, expand_subject_hierarchy
 from osf.utils.fields import NonNaiveDateTimeField
 from osf.utils.datetime_aware_jsonfield import DateTimeAwareJSONField
-from osf.utils.machines import ReviewsMachine, NodeRequestMachine, PreprintRequestMachine
+from osf.utils.machines import (
+    ReviewsMachine,
+    NodeRequestMachine,
+    PreprintRequestMachine,
+    RegistrationMachine
+)
+
 from osf.utils.permissions import ADMIN, REVIEW_GROUPS, READ, WRITE
 from osf.utils.registrations import flatten_registration_metadata, expand_registration_responses
-from osf.utils.workflows import DefaultStates, DefaultTriggers, ReviewStates, ReviewTriggers
+from osf.utils.workflows import (
+    DefaultStates,
+    DefaultTriggers,
+    ReviewStates,
+    ReviewTriggers,
+    RegistrationStates,
+    RegistrationTriggers
+)
+
 from osf.utils.requests import get_request_and_user_id
 from website.project import signals as project_signals
 from website import settings, mails, language
@@ -2228,3 +2242,68 @@ class EditableFieldsMixin(TitleMixin, DescriptionMixin, CategoryMixin, Contribut
 
     class Meta:
         abstract = True
+
+
+class RegistriesModerationMixin(MachineableMixin):
+    """This is to facilitate the registraies modeation process and should extend the DraftRegistration model.
+    """
+    TriggersClass = RegistrationTriggers
+
+    machine_state = models.CharField(
+        max_length=30,
+        db_index=True,
+        choices=RegistrationStates.choices(),
+        default=RegistrationStates.INITIAL.value
+    )
+
+    class Meta:
+        abstract = True
+
+    MachineClass = RegistrationMachine
+
+    def run_request_withdraw(self, user, comment):
+        """Run the 'request_withdraw' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        return self._run_transition(self.TriggersClass.REQUEST_WITHDRAW.value, user=user, comment=comment)
+
+    def run_withdraw_registration(self, user, comment):
+        """Run the 'withdraw' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        return self._run_transition(self.TriggersClass.WITHDRAW.value, user=user, comment=comment)
+
+    def run_accept(self, user, comment, embargo_end_date=None):
+        """ Run the 'accept/embargo' state transition and create a corresponding Action.
+
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        if embargo_end_date:
+            return self._run_transition(self.TriggersClass.EMBARGO.value, user=user, end_date=embargo_end_date)
+        else:
+            return self._run_transition(self.TriggersClass.ACCEPT.value, user=user)
+
+    def run_request_embargo_termination(self, user, comment):
+        """Run the 'embargo_termination' state transition and create a corresponding Action.
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+
+        return self._run_transition(self.TriggersClass.REQUEST_EMBARGO_TERMINATION.value, user=user, comment=comment)
+
+    def run_terminate_embargo(self, user, comment):
+        """Run the 'embargo_termination' state transition and create a corresponding Action.
+        Params:
+            user: The user triggering this transition.
+            comment: Text describing why.
+        """
+        return self._run_transition(self.TriggersClass.TERMINATE_EMBARGO.value, user=user, comment=comment)
