+++ osf/models/registrations.py
@@ -42,16 +42,14 @@ from osf.models.mixins import (
     EditableFieldsMixin,
     Loggable,
     GuardianMixin,
-    MachineableMixin
+    RegistriesModerationMixin
 )
 from osf.models.nodelog import NodeLog
 from osf.models.provider import RegistrationProvider
 from osf.models.mixins import RegistrationResponseMixin
-from osf.utils.machines import RegistrationMachine
 from osf.models.tag import Tag
 from osf.models.validators import validate_title
 from osf.utils.datetime_aware_jsonfield import DateTimeAwareJSONField
-from osf.utils.workflows import RegistrationTriggers, RegistrationStates
 
 logger = logging.getLogger(__name__)
 
@@ -566,9 +564,7 @@ def get_default_id():
 
 
 class DraftRegistration(ObjectIDMixin, RegistrationResponseMixin, DirtyFieldsMixin,
-        BaseModel, Loggable, EditableFieldsMixin, GuardianMixin, MachineableMixin):
-
-    TriggersClass = RegistrationTriggers
+        BaseModel, Loggable, EditableFieldsMixin, GuardianMixin, RegistriesModerationMixin):
 
     # Fields that are writable by DraftRegistration.update
     WRITABLE_WHITELIST = [
@@ -577,10 +573,6 @@ class DraftRegistration(ObjectIDMixin, RegistrationResponseMixin, DirtyFieldsMix
         'category',
         'node_license',
     ]
-
-    @property
-    def MachineClass(self):
-        return RegistrationMachine
 
     URL_TEMPLATE = settings.DOMAIN + 'project/{node_id}/drafts/{draft_id}'
 
@@ -655,13 +647,6 @@ class DraftRegistration(ObjectIDMixin, RegistrationResponseMixin, DirtyFieldsMix
         'admin': (READ_DRAFT_REGISTRATION, WRITE_DRAFT_REGISTRATION, ADMIN_DRAFT_REGISTRATION,)
     }
     group_format = 'draft_registration_{self.id}_{group}'
-
-    machine_state = models.CharField(
-        max_length=30,
-        db_index=True,
-        choices=RegistrationStates.choices(),
-        default=RegistrationStates.INITIAL.value
-    )
 
     class Meta:
         permissions = (
@@ -1153,39 +1138,6 @@ class DraftRegistration(ObjectIDMixin, RegistrationResponseMixin, DirtyFieldsMix
             self.save()
 
         return updated
-
-    def run_request_withdraw(self, user, comment):
-        """Run the 'request_withdraw' state transition and create a corresponding Action.
-
-        Params:
-            user: The user triggering this transition.
-            comment: Text describing why.
-        """
-        return self._run_transition(self.TriggersClass.REQUEST_WITHDRAW.value, user=user, comment=comment)
-
-    def run_withdraw_registration(self, user, comment):
-        """Run the 'withdraw' state transition and create a corresponding Action.
-
-        Params:
-            user: The user triggering this transition.
-            comment: Text describing why.
-        """
-        return self._run_transition(self.TriggersClass.WITHDRAW.value, user=user, comment=comment)
-
-    def run_accept(self, user, comment, embargo_end_date=None):
-        """ Run the 'accept/embargo' state transition and create a corresponding Action. """
-        if embargo_end_date:
-            return self._run_transition(self.TriggersClass.EMBARGO.value, user=user, end_date=embargo_end_date)
-        else:
-            return self._run_transition(self.TriggersClass.ACCEPT.value, user=user)
-
-    def run_request_embargo_termination(self, user, comment):
-        """Run the 'embargo_termination' state transition and create a corresponding Action. """
-        return self._run_transition(self.TriggersClass.REQUEST_EMBARGO_TERMINATION.value, user=user, comment=comment)
-
-    def run_terminate_embargo(self, user, comment):
-        """Run the 'embargo_termination' state transition and create a corresponding Action. """
-        return self._run_transition(self.TriggersClass.TERMINATE_EMBARGO.value, user=user, comment=comment)
 
 
 class DraftRegistrationUserObjectPermission(UserObjectPermissionBase):
